Export actually works ... no idea why
